requirements.txt
# Backend dependencies
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.5
python-dotenv==1.0.0

# Frontend dependencies will be managed separately with package.json